requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==5.1
django-allauth==64.2.0
django-cors-headers==4.4.0
django-filter==24.3
django-taggit==6.0.0
djangorestframework==3.15.2
google-ai-generativelanguage==0.6.15
google-api-core==2.19.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.65.0
grpcio==1.70.0
grpcio-status==1.62.3
gunicorn==23.0.0
httplib2==0.22.0
idna==3.8
Markdown==3.7
packaging==24.1
proto-plus==1.26.0
protobuf==4.25.4
psycopg==3.2.5
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.4
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
